adecuacion a modelo VMMV
