+++ app/src/main/res/layout/fragment_login.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/dark_gray"
-    tools:context=".view.fragment.LoginActivity">
+    tools:context=".view.fragment.LoginFragment">
 
     <ImageView
         android:id="@+id/background_image"
