+++ app/src/main/res/navigation/navigation_graph.xml
@@ -3,9 +3,52 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/navigation_graph"
-    app:startDestination="@id/detailAppointmentFragment">
+    app:startDestination="@id/loginFragment">
 
+    <fragment
+        android:id="@+id/loginFragment"
+        android:name="com.example.miniproyecto_i.view.fragment.LoginFragment"
+        android:label="LoginFragment"
+        tools:layout="@layout/fragment_login">
+        <action
+            android:id="@+id/action_loginFragment_to_homeAppointments"
+            app:destination="@id/homeAppointments"
+            app:popUpTo="@id/loginFragment"
+            app:popUpToInclusive="true" />
+    </fragment>
 
+    <fragment
+        android:id="@+id/homeAppointments"
+        android:name="com.example.miniproyecto_i.view.fragment.HomeAppointments"
+        android:label="HomeAppointments" >
+        <action
+            android:id="@+id/action_homeAppointments_to_detailAppointmentFragment"
+            app:destination="@id/detailAppointmentFragment">
+            <argument
+                android:name="appointment"
+                app:argType="com.example.miniproyecto_i.model.Appointment" />
+            <argument
+                android:name="turno"
+                app:argType="integer" />
+        </action>
+
+        <action
+            android:id="@+id/action_homeAppointments_to_addAppointmentFragment"
+            app:destination="@id/addAppointmentFragment" />
+    </fragment>
+
+    <fragment
+        android:id="@+id/detailAppointmentFragment"
+        android:name="com.example.miniproyecto_i.view.fragment.DetailAppointmentFragment"
+        android:label="DetailAppointmentFragment"
+        tools:layout="@layout/fragment_detail_appointment">
+        <action
+            android:id="@+id/action_itemDetailsFragment_to_itemEditFragment"
+            app:destination="@id/editAppointmentFragment" />
+        <action
+            android:id="@+id/action_itemDetailsFragment_to_itemAddFragment"
+            app:destination="@id/addAppointmentFragment" />
+    </fragment>
 
     <fragment
         android:id="@+id/editAppointmentFragment"
@@ -18,30 +61,6 @@
     </fragment>
 
     <fragment
-        android:id="@+id/detailAppointmentFragment"
-        android:name="com.example.miniproyecto_i.view.fragment.DetailAppointmentFragment"
-        android:label="detailAppointmentFragment"
-        tools:layout="@layout/fragment_detail_appointment">
-        <action
-            android:id="@+id/action_itemDetailsFragment_to_itemEditFragment"
-            app:destination="@id/editAppointmentFragment" />
-        <action
-            android:id="@+id/action_itemDetailsFragment_to_itemAddFragment"
-            app:destination="@id/addAppointmentFragment" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/homeAppointments"
-        android:name="com.example.miniproyecto_i.view.fragment.HomeAppointments"
-        android:label="HomeAppointments" >
-        <action
-            android:id="@+id/action_homeAppointments_to_detailAppointmentFragment"
-            app:destination="@id/detailAppointmentFragment" />
-        <action
-            android:id="@+id/action_homeAppointments_to_addAppointmentFragment"
-            app:destination="@id/addAppointmentFragment" />
-    </fragment>
-    <fragment
         android:id="@+id/addAppointmentFragment"
         android:name="com.example.miniproyecto_i.view.fragment.AddAppointmentFragment"
         android:label="AddAppointmentFragment" >
@@ -49,5 +68,4 @@
             android:id="@+id/action_addAppointmentFragment_to_homeAppointments"
             app:destination="@id/homeAppointments" />
     </fragment>
-
 </navigation>
